Web/joke-api/deployment/requirements.txt
Flask
pyjwt==1.5.0
cryptography==2.6.1